reset the repository (mostly)
contribute.json can mostly be rewritten as a static site to reduce maintenance costs. Rather than start from an orphan branch we'll preserve the history and contributors and whatnot and simply strip the repository down to get started.

Some functionality may not be possible as a static site only and may be lost. Let's see how far we can get.